# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DealBrief AI API
  version: 1.0.0
  description: |
    Flask and PostgreSQL backend for customer product subscriptions,
    AI-generated customer intelligence, and meeting briefs.

    Successful responses use `{ "data": ... }`. Error responses use
    `{ "error": { "code": "...", "message": "...", "details": null } }`.
servers:
  - url: /api
    description: Current Flask server
  - url: http://127.0.0.1:3000/api
    description: Local development server
...
security:
  - sessionCookie: []
paths:
...
  /auth/register:
    post:
      security: []
      tags: [Authentication]
      summary: Register a user account
      description: All self-registered accounts receive the `user` role.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: {type: string, maxLength: 100}
                email: {type: string, format: email}
                password: {type: string, minLength: 8}
      responses:
        "201": {description: Account created and session started.}
        "409": {description: Email is already registered.}
        "422": {$ref: "#/components/responses/ValidationError"}
  /auth/login:
    post:
      security: []
      tags: [Authentication]
      summary: Sign in with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string}
      responses:
        "200": {description: Session started.}
        "401": {description: Email or password is incorrect.}
...
  /auth/me:
    get:
      tags: [Authentication]
      summary: Return the signed-in user
      operationId: getCurrentUser
      responses:
        "200": {description: Current user profile and role.}
        "401": {description: Authentication is required.}
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      operationId: listCustomers
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, pilot, inactive]
          example: active
        - name: search
          in: query
          required: false
          description: Case-insensitive partial match against customer name.
          schema:
            type: string
          example: Bank
      responses:
        "200":
          description: Customers that have not been soft deleted.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Customer"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Customers]
      summary: Add a customer
      description: Administrator only.
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, maxLength: 200}
                industry: {type: string}
      responses:
        "201": {description: Customer created.}
        "403": {description: Administrator access is required.}
        "422": {$ref: "#/components/responses/ValidationError"}
  /customers/{customerId}:
    delete:
      tags: [Customers]
      summary: Soft delete a customer
      operationId: deleteCustomer
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Customer was marked inactive and assigned a deleted timestamp.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
              example:
                data:
                  id: 30000000-0000-0000-0000-000000000001
                  deleted: true
                  deleteMode: soft
                  deletedAt: "2026-07-31T18:45:00+00:00"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /customers/{customerId}/dashboard:
    get:
      tags: [Customers]
      summary: Get the customer intelligence dashboard
      operationId: getCustomerDashboard
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Customer profile, latest subscriptions by product, and latest intelligence.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/CustomerDashboard"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /customers/{customerId}/timeline:
    get:
      tags: [Subscriptions]
      summary: Get licensed seats across subscription periods
      operationId: getCustomerTimeline
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: One licensed-seat series per product, combining annual renewal records.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/SubscriptionTimeline"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /products:
    get:
      tags: [Products]
      summary: List active products
      operationId: listProducts
      responses:
        "200":
          description: Active products.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Products]
      summary: Add a product
      description: Administrator only.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, maxLength: 150}
                description: {type: string}
      responses:
        "201": {description: Product created.}
        "403": {description: Administrator access is required.}
        "409": {description: Product name already exists.}
  /subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      operationId: listSubscriptions
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          example: 30000000-0000-0000-0000-000000000001
      responses:
        "200":
          description: Up to 500 subscriptions ordered by start date descending.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subscription"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [Subscriptions]
      summary: Add a customer product subscription
      description: Administrator only.
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerId, productId, subscriptionStartDate]
              properties:
                customerId: {type: string, format: uuid}
                productId: {type: string, format: uuid}
                subscriptionStartDate: {type: string, format: date}
                subscriptionEndDate: {type: string, format: date}
                subscriptionStatus:
                  type: string
                  enum: [active, expired, canceled]
                licensedSeats: {type: integer, minimum: 0}
      responses:
        "201": {description: Subscription created.}
        "403": {description: Administrator access is required.}
        "409": {description: The subscription already exists for this customer, product, and start date.}
        "422": {$ref: "#/components/responses/ValidationError"}
  /subscriptions/{subscriptionId}:
    delete:
      tags: [Subscriptions]
      summary: Permanently delete a subscription
      operationId: deleteSubscription
      parameters:
        - $ref: "#/components/parameters/SubscriptionId"
      responses:
        "200":
          description: Subscription was hard deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
              example:
                data:
                  id: 40000000-0000-0000-0000-000000000001
                  deleted: true
                  deleteMode: hard
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /customers/{customerId}/intelligence:
    get:
      tags: [Intelligence]
      summary: List intelligence snapshots
      operationId: listIntelligence
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Up to 50 intelligence snapshots ordered newest first.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/IntelligenceSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /customers/{customerId}/intelligence/latest:
    get:
      tags: [Intelligence]
      summary: Get the latest intelligence snapshot
      operationId: getLatestIntelligence
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Latest intelligence snapshot.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/IntelligenceSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /customers/{customerId}/intelligence/refresh:
    post:
      tags: [Intelligence]
      summary: Generate and save a new AI intelligence snapshot
      description: |
        Uses stored subscription records, calls OpenAI, validates
        the structured result, and saves it to PostgreSQL.
      operationId: refreshIntelligence
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntelligenceRefreshRequest"
            example:
              periodStart: "2026-05-01"
              periodEnd: "2026-07-31"
      responses:
        "201":
          description: New intelligence snapshot.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/IntelligenceSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: No subscription exists in the requested period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /generate-brief:
    post:
      tags: [Meeting Briefs]
      summary: Generate an AI meeting brief
      description: Calls OpenAI and validates the structured output without saving a log.
      operationId: generateMeetingBrief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBriefRequest"
            example:
              customerId: 30000000-0000-0000-0000-000000000001
              productId: 20000000-0000-0000-0000-000000000001
              meetingType: winback
              deliverableType: call_brief
              notes: Focus on renewal value and reporting adoption.
      responses:
        "201":
          description: Generated brief.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/GeneratedBrief"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 30000000-0000-0000-0000-000000000001
    SubscriptionId:
      name: subscriptionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 40000000-0000-0000-0000-000000000001
  responses:
    ValidationError:
      description: Request parameters or JSON are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: customerId must be a valid UUID.
              details: null
    NotFound:
      description: The requested database object does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
              details: null
    InternalError:
      description: Unexpected server or database error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
    Customer:
      type: object
      required:
        - id
        - name
        - industry
        - status
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        industry:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pilot, inactive]
      example:
        id: 30000000-0000-0000-0000-000000000001
        name: ABC Bank
        industry: Financial Services
        status: active
    Product:
      type: object
      required: [id, name, description, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
      example:
        id: 20000000-0000-0000-0000-000000000001
        name: CRM Platform
        description: Pipeline, account, and sales workflow management.
        status: active
    Subscription:
      type: object
      required:
        - id
        - customerId
        - customerName
        - productId
        - productName
        - subscriptionStartDate
        - subscriptionStatus
        - licensedSeats
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        productId:
          type: string
          format: uuid
        productName:
          type: string
        subscriptionStartDate:
          type: string
          format: date
        subscriptionEndDate:
          type: string
          format: date
          nullable: true
        subscriptionStatus:
          type: string
          enum: [active, expired, canceled]
        licensedSeats:
          type: integer
          minimum: 0
          nullable: true
      example:
        id: 40000000-0000-0000-0000-000000000001
        customerId: 30000000-0000-0000-0000-000000000001
        customerName: ABC Bank
        productId: 20000000-0000-0000-0000-000000000001
        productName: CRM Platform
        subscriptionStartDate: "2026-01-01"
        subscriptionEndDate: "2026-12-31"
        subscriptionStatus: active
        licensedSeats: 500
...
    IndustryDynamic:
      type: object
      required: [headline, summary, impact]
      properties:
        headline: {type: string}
        summary: {type: string}
        impact: {type: string}
...
    IntelligenceSnapshot:
      type: object
      required:
        - id
        - aiKeySignal
        - industryDynamics
        - companyNews
        - recommendedNextSteps
        - metrics
        - generatedAt
      properties:
        id:
          type: string
          format: uuid
        aiKeySignal:
          type: string
        industryDynamics:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: "#/components/schemas/IndustryDynamic"
        companyNews:
          type: array
          items:
            $ref: "#/components/schemas/CompanyNewsItem"
        recommendedNextSteps:
          $ref: "#/components/schemas/RecommendedNextSteps"
        metrics:
          $ref: "#/components/schemas/IntelligenceMetrics"
        generatedAt:
          type: string
          format: date-time
    CustomerDashboard:
      type: object
      required: [customer, latestSubscriptions, intelligence]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        latestSubscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        intelligence:
          allOf:
            - $ref: "#/components/schemas/IntelligenceSnapshot"
          nullable: true
...
    SubscriptionTimeline:
      type: object
      required: [customer, periodStart, periodEnd, series]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        periodStart:
          type: string
          format: date
          nullable: true
        periodEnd:
          type: string
          format: date
          nullable: true
        series:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionTimelineSeries"
    IntelligenceRefreshRequest:
      type: object
      properties:
        periodStart:
          type: string
          description: ISO 8601 date or timestamp; defaults to 90 days before periodEnd.
          example: "2026-05-01"
        periodEnd:
          type: string
          description: ISO 8601 date or timestamp; defaults to the current time.
          example: "2026-07-31"
...
    DeleteResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, deleted, deleteMode]
          properties:
            id:
              type: string
              format: uuid
            deleted:
              type: boolean
            deleteMode:
              type: string
              enum: [soft, hard]
            deletedAt:
              type: string
              format: date-time
            status:
              type: string